changed directives function to export a function
